services/spread-sheet/requirements.txt
Flask==3.1.1
flask_sqlalchemy==3.1.1
gspread==6.2.1
langchain_openai==0.3.19
matplotlib==3.10.3
openpyxl==3.1.5
pandas==1.5.3
numpy==1.26.4
protobuf==6.31.1
psycopg2_binary==2.9.9
python-dotenv==1.1.0
